requirements.txt
streamlit
pinecone-client
openai
sentence-transformers
python-dotenv
